requirements.txt
fastapi>=0.115.12
granian>=2.2.6
httpx>=0.28.1
pydantic-settings>=2.9.1
pydantic>=2.11.4
pyyaml>=6.0.2
sse-starlette>=2.3.5
uvicorn==0.34.2
